utils: move Size to its own submodule
