+++ src/utils/mod.rs
@@ -1,16 +1,11 @@
 use errors::*;
-use serde::{
-    de::{Deserialize, Deserializer, Error as DeError, Visitor},
-    ser::{Serialize, Serializer},
-};
 use std::any::Any;
-use std::fmt;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::result::Result as StdResult;
-use std::str::FromStr;
 use std::thread;
 use std::time::Duration;
+
+pub mod size;
 
 #[macro_use]
 macro_rules! string_enum {
@@ -207,79 +202,6 @@ pub fn from_hex(input: &str) -> Result<Vec<u8>> {
     }
 }
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum Size {
-    Bytes(usize),
-    Kilobytes(usize),
-    Megabytes(usize),
-    Gigabytes(usize),
-    Terabytes(usize),
-}
-
-impl fmt::Display for Size {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Size::Bytes(count) => write!(f, "{}", count),
-            Size::Kilobytes(count) => write!(f, "{}K", count),
-            Size::Megabytes(count) => write!(f, "{}M", count),
-            Size::Gigabytes(count) => write!(f, "{}G", count),
-            Size::Terabytes(count) => write!(f, "{}T", count),
-        }
-    }
-}
-
-impl FromStr for Size {
-    type Err = Error;
-
-    fn from_str(mut input: &str) -> Result<Size> {
-        let mut last = input.chars().last().ok_or("empty size")?;
-
-        // Eat a trailing 'b'
-        if last == 'b' || last == 'B' {
-            input = &input[..input.len() - 1];
-            last = input.chars().last().ok_or("empty size")?;
-        }
-
-        if last == 'K' || last == 'k' {
-            Ok(Size::Kilobytes(input[..input.len() - 1].parse()?))
-        } else if last == 'M' || last == 'm' {
-            Ok(Size::Megabytes(input[..input.len() - 1].parse()?))
-        } else if last == 'G' || last == 'g' {
-            Ok(Size::Gigabytes(input[..input.len() - 1].parse()?))
-        } else if last == 'T' || last == 't' {
-            Ok(Size::Terabytes(input[..input.len() - 1].parse()?))
-        } else {
-            Ok(Size::Bytes(input.parse()?))
-        }
-    }
-}
-
-struct SizeVisitor;
-
-impl<'de> Visitor<'de> for SizeVisitor {
-    type Value = Size;
-
-    fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str("a size")
-    }
-
-    fn visit_str<E: DeError>(self, input: &str) -> StdResult<Size, E> {
-        Size::from_str(input).map_err(E::custom)
-    }
-}
-
-impl<'de> Deserialize<'de> for Size {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> StdResult<Size, D::Error> {
-        deserializer.deserialize_str(SizeVisitor)
-    }
-}
-
-impl Serialize for Size {
-    fn serialize<S: Serializer>(&self, serializer: S) -> StdResult<S::Ok, S::Error> {
-        serializer.serialize_str(&self.to_string())
-    }
-}
-
 pub fn split_quoted(input: &str) -> Result<Vec<String>> {
     let mut segments = Vec::new();
     let mut buffer = String::new();
@@ -317,39 +239,7 @@ pub fn split_quoted(input: &str) -> Result<Vec<String>> {
 
 #[cfg(test)]
 mod tests {
-    use super::{split_quoted, Size};
-
-    #[test]
-    fn test_size() {
-        assert_eq!("1234".parse::<Size>().unwrap(), Size::Bytes(1234));
-        assert_eq!("1234B".parse::<Size>().unwrap(), Size::Bytes(1234));
-        assert_eq!("1234b".parse::<Size>().unwrap(), Size::Bytes(1234));
-        assert_eq!(Size::Bytes(1234).to_string(), "1234");
-
-        assert_eq!("1234K".parse::<Size>().unwrap(), Size::Kilobytes(1234));
-        assert_eq!("1234k".parse::<Size>().unwrap(), Size::Kilobytes(1234));
-        assert_eq!("1234KB".parse::<Size>().unwrap(), Size::Kilobytes(1234));
-        assert_eq!("1234kb".parse::<Size>().unwrap(), Size::Kilobytes(1234));
-        assert_eq!(Size::Kilobytes(1234).to_string(), "1234K");
-
-        assert_eq!("1234M".parse::<Size>().unwrap(), Size::Megabytes(1234));
-        assert_eq!("1234m".parse::<Size>().unwrap(), Size::Megabytes(1234));
-        assert_eq!("1234MB".parse::<Size>().unwrap(), Size::Megabytes(1234));
-        assert_eq!("1234mb".parse::<Size>().unwrap(), Size::Megabytes(1234));
-        assert_eq!(Size::Megabytes(1234).to_string(), "1234M");
-
-        assert_eq!("1234G".parse::<Size>().unwrap(), Size::Gigabytes(1234));
-        assert_eq!("1234g".parse::<Size>().unwrap(), Size::Gigabytes(1234));
-        assert_eq!("1234GB".parse::<Size>().unwrap(), Size::Gigabytes(1234));
-        assert_eq!("1234Gb".parse::<Size>().unwrap(), Size::Gigabytes(1234));
-        assert_eq!(Size::Gigabytes(1234).to_string(), "1234G");
-
-        assert_eq!("1234T".parse::<Size>().unwrap(), Size::Terabytes(1234));
-        assert_eq!("1234t".parse::<Size>().unwrap(), Size::Terabytes(1234));
-        assert_eq!("1234TB".parse::<Size>().unwrap(), Size::Terabytes(1234));
-        assert_eq!("1234Tb".parse::<Size>().unwrap(), Size::Terabytes(1234));
-        assert_eq!(Size::Terabytes(1234).to_string(), "1234T");
-    }
+    use super::split_quoted;
 
     #[test]
     fn test_split_quoted() {
